can view and add subtasks from 'Current Task'
