+++ src/components/current-task/CurrentTask.jsx
@@ -24,7 +24,8 @@ import {
  * with a timer that can be used to complete the current task.
  * Timer configuration is available to the user.
  */
-function CurrentTask() {
+function CurrentTask(props) {
+  const { subtasks, onNewSubtask } = props;
   const [currentTask] = useContext(CurrentTaskContext);
   const [, setWorkTimerMemory] = useContext(WorkTimerMemoryContext);
   const [, setBreakTimerMemory] = useContext(BreakTimerMemoryContext);
@@ -116,7 +117,9 @@ function CurrentTask() {
               <Task
                 expanded
                 name={currentTask.Name}
-                subtasks={currentTask.Subtasks}
+                subtasks={subtasks.filter(
+                  (subtask) => subtask.Task_id === currentTask.Task_id
+                )}
                 endDate={currentTask.End_date}
                 label={currentTask.Label}
                 startDate={currentTask.Start_date}
@@ -124,6 +127,7 @@ function CurrentTask() {
                 checked={currentTask.Is_complete}
                 description={currentTask.Description}
                 userId={currentTask.User_id}
+                onNewSubtask={onNewSubtask}
               />
               <CurrentTaskNotes
                 notes={[currentTask.Description]}
