+++ src/components/global/TaskList.jsx
@@ -1,11 +1,9 @@
-import React, { useContext, useEffect, useState } from "react";
+import React, { useContext } from "react";
 import moment from "moment";
 import NewTask from "../NewTask";
 import Task from "./Task/Task";
 import { CurrentTaskContext } from "../timer-modal/TimerContextProvider";
 import DarkModeContext from "../../services/theme-context";
-import { AuthContext } from "../../services/providers/authProvider";
-import { fetchSubtasks } from "../../services/databaseService";
 
 /**
  * This function represents the global Task List component used on the home page
@@ -13,18 +11,9 @@ import { fetchSubtasks } from "../../services/databaseService";
  * display task details, and a button to add new tasks.
  */
 function TaskList(props) {
-  const { onTaskClick, tasks, onNewTask } = props;
+  const { onTaskClick, tasks, subtasks, onNewTask, onNewSubtask } = props;
   const [currentTask] = useContext(CurrentTaskContext);
   const { isDarkMode } = React.useContext(DarkModeContext);
-  const { currentUser } = useContext(AuthContext);
-  const [subtasks, setSubtasks] = useState([]);
-  const [refetchSubtasks, setRefetchSubtasks] = useState(false);
-
-  const triggerRefetchSubtasks = () => setRefetchSubtasks(!refetchSubtasks);
-
-  useEffect(() => {
-    fetchSubtasks(currentUser.uid).then((res) => setSubtasks(res));
-  }, [refetchSubtasks]);
 
   const today = new Date().toDateString();
   const seconds = moment(today, false).toDate().getTime();
@@ -56,7 +45,7 @@ function TaskList(props) {
                     subtasks={subtasks.filter(
                       (subtask) => subtask.Task_id === t.Task_id
                     )}
-                    onNewSubtask={triggerRefetchSubtasks}
+                    onNewSubtask={onNewSubtask}
                   />
                 )}
             </div>
