+++ src/pages/dashboardView.js
@@ -4,8 +4,11 @@ import TaskList from "../components/global/TaskList";
 import GoalList from "../components/goal/GoalList";
 import Navbar from "../components/Navbar";
 import { CurrentTaskContext } from "../components/timer-modal/TimerContextProvider";
-
-import { fetchGoals, fetchTasks } from "../services/databaseService";
+import {
+  fetchGoals,
+  fetchTasks,
+  fetchSubtasks,
+} from "../services/databaseService";
 import { AuthContext } from "../services/providers/authProvider";
 import DarkModeContext from "../services/theme-context";
 
@@ -31,10 +34,12 @@ function DashboardView() {
   const { currentUser } = useContext(AuthContext);
   const [, setCurrentTask] = useContext(CurrentTaskContext);
   const [tasks, setTasks] = useState([]);
+  const [subtasks, setSubtasks] = useState([]);
   const [goals, setGoals] = useState([]);
   const { isDarkMode } = React.useContext(DarkModeContext);
   const [refetchTasks, setRefetchTasks] = useState(false);
   const [refetchGoals, setRefetchGoals] = useState(false);
+  const [refetchSubtasks, setRefetchSubtasks] = useState(false);
 
   /**
    * When a task is selected from Today's Tasks,
@@ -52,6 +57,8 @@ function DashboardView() {
     setRefetchGoals(!refetchGoals);
   };
 
+  const triggerRefetchSubtasks = () => setRefetchSubtasks(!refetchSubtasks);
+
   useEffect(() => {
     fetchTasks(currentUser.uid).then((res) => {
       setTasks(res);
@@ -59,7 +66,10 @@ function DashboardView() {
     fetchGoals(currentUser.uid).then((res) => {
       setGoals(res);
     });
-  }, [refetchTasks, refetchGoals]);
+    fetchSubtasks(currentUser.uid).then((res) => {
+      setSubtasks(res);
+    });
+  }, [refetchTasks, refetchGoals, refetchSubtasks]);
 
   return (
     <>
@@ -67,10 +77,15 @@ function DashboardView() {
       <div className={isDarkMode ? "dashboard" : "dashboard light"}>
         <TaskList
           tasks={tasks}
+          subtasks={subtasks}
           onNewTask={triggerRefetchTasks}
           onTaskClick={handleTaskClick}
+          onNewSubtask={triggerRefetchSubtasks}
         />
-        <CurrentTask />
+        <CurrentTask
+          subtasks={subtasks}
+          onNewSubtask={triggerRefetchSubtasks}
+        />
         <div className="dashboard__placeholder-column">
           <DashboardPlaceholder title="Stats:" />
           <br />
