+++ src/stylesheets/components/global/task/task.scss
@@ -50,4 +50,9 @@
 	&__content--expanded {
 		overflow: auto;
 	}
+
+	&__new-subtask {
+		display: flex;
+		justify-content: center;
+	}
 }
